Move validation functionality into module
